Add LCNames Authority
Adds support for LCNames, a custom authority with a default dataset
name. We use the existing search logic that ships with QA. To support
this we add a search service interface and test compliance for the
existing Qa LOC support.

Makes tests for `LDF::Authority` sharable by making them shared examples
and moving them to `lib/qa/ldf/spec`.